Remove email templates and route
Although I want to give admins the ability to save and edit the message
they send to bidders at the end of an auction, I want to assess the
best way to do this. As such, I don’t want to add the table in the
database until this is determined.